% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{countHeaderLines}
\alias{countHeaderLines}
\title{Count the number of lines at the beginning of the file that start with @ or #}
\usage{
countHeaderLines(path)
}
\arguments{
\item{path}{text file to be read, possibly gzipped}
}
\value{
the number of lines at the beginning of the file that start with '@' or '#'
}
\description{
Count the number of lines at the beginning of the file that start with @ or #
}
