% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{datatable.cedata}
\alias{datatable.cedata}
\title{Report if data.table is set up incorrectly by temporarily turning on the option datatable.verbose and calling the
internal function data.table:::cedta().}
\usage{
datatable.cedata(n = 3L)
}
\arguments{
\item{n}{parameter to pass to `data.table:::cedta()`}
}
\value{
the result of calling `data.table:::cedta(n)`
}
\description{
With verbosity turned on, that private function warns about a case where if the data.table package is not listed in
a package NAMESPACE file then any data.table indexing will run VERY slowly as `[.data.table` will revert to calling
`[.data.frame`.
}
\details{
See also:
- https://github.com/Rdatatable/data.table/blob/1.17.6/R/cedta.R#L42-L43
- https://github.com/Rdatatable/data.table/blob/1.17.6/vignettes/datatable-importing.Rmd#L189
- https://stackoverflow.com/questions/10527072/using-data-table-package-inside-my-own-package#answer-10529888
- https://github.com/Rdatatable/data.table/issues/2341
- https://github.com/Rdatatable/data.table/issues/3744
- etc.
}
