% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{ensure_trailing_slash}
\alias{ensure_trailing_slash}
\title{Make sure the argument has a trailing slash.}
\usage{
ensure_trailing_slash(directory)
}
\arguments{
\item{directory}{path on which to operate}
}
\value{
the input argument, with a slash appended if it doesn't already end with a slash
}
\description{
Make sure the argument has a trailing slash.
}
