% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{fastRead}
\alias{fastRead}
\title{Read a data.table using data.table fread()}
\usage{
fastRead(inFile, comment_regexp = NA, ...)
}
\arguments{
\item{inFile}{path of file to be read}

\item{comment_regexp}{If defined, filter input through "egrep -v comment_regexp"}

\item{...}{passed through to data.table::fread}
}
\value{
a data.table
}
\description{
If inFile ends with .gz, it is gunzipped in a way that is robust to running out of disk space.
Environment variable DROPSEQ_TMPDIR should be set to a directory that has plenty of room and is writable.
}
