% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{getDenseDgeDimensions}
\alias{getDenseDgeDimensions}
\title{Read enough of a file to get the number of rows and columns.}
\usage{
getDenseDgeDimensions(file)
}
\arguments{
\item{file}{possibly gzipped}
}
\value{
list(numRows, numCols)
}
\description{
Read enough of a file to get the number of rows and columns.
}
