% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{getDgeDimensions}
\alias{getDgeDimensions}
\title{Determine size of sparse or dense DGE}
\usage{
getDgeDimensions(file)
}
\arguments{
\item{file}{possibly gzipped}
}
\value{
list(numRows, numCols, numPopulatedElements)
}
\description{
Determine size of sparse or dense DGE
}
