% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{getSparseMatrixDimensions}
\alias{getSparseMatrixDimensions}
\title{read the first fiew lines of a MatrixMarket file to get the header line.}
\usage{
getSparseMatrixDimensions(file)
}
\arguments{
\item{file}{possibly gzipped}
}
\value{
list(numRows, numCols, numPopulatedElements)
}
\description{
read the first fiew lines of a MatrixMarket file to get the header line.
}
