% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{isMatrixMarket}
\alias{isMatrixMarket}
\title{Determine if a file is in matrix market file format.}
\usage{
isMatrixMarket(file)
}
\arguments{
\item{file}{The file to test}
}
\value{
true if the file is in matrix market format
}
\description{
Determine if a file is in matrix market file format.
}
