% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\docType{data}
\name{MatrixMarketConstants}
\alias{MatrixMarketConstants}
\title{Strings that appear in MatrixMarket files.}
\format{
An object of class \code{list} of length 9.
}
\usage{
MatrixMarketConstants
}
\description{
Strings that appear in MatrixMarket files.
}
\keyword{datasets}
