% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_util.R
\name{maybeRemoveSuffix}
\alias{maybeRemoveSuffix}
\title{Remove a suffix from a string, if it is present.}
\usage{
maybeRemoveSuffix(str, suffix)
}
\arguments{
\item{str}{the string to be truncated}

\item{suffix}{the string to be checked for and removed}
}
\value{
the string with suffix removed, if present, otherwise the original string
}
\description{
If the string does not end with the suffix, it is not truncated.
}
