% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{read_dge_gz}
\alias{read_dge_gz}
\title{Read a tabular DGE, with gene rows and columns that typically are cell barcodes}
\usage{
read_dge_gz(file, decreasing_order_by_size = TRUE)
}
\arguments{
\item{file}{to be read, optionally gzipped}

\item{decreasing_order_by_size}{If true, columns are ordered in decreasing order by colsum}
}
\value{
a data.frame with genes as rownames, no GENE column.
}
\description{
Read a tabular DGE, with gene rows and columns that typically are cell barcodes
}
