% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{read_dge_txt}
\alias{read_dge_txt}
\title{Read a tabular DGE, that may have comment/header lines starting with '@' or '#'}
\usage{
read_dge_txt(path)
}
\arguments{
\item{path}{file to be read.  If has suffix ".gz" it is assumed to be gzipped.}
}
\value{
a data.table with the tabular contents of the file, and header/comment lines attached
        as an attribute named 'Drop-seq.dge.header'
}
\description{
Read a tabular DGE, that may have comment/header lines starting with '@' or '#'
}
