% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{read_table_with_label_for_rownames}
\alias{read_table_with_label_for_rownames}
\title{Read a tab-separated file, drop the first column and use the values of the first column as rownames.}
\usage{
read_table_with_label_for_rownames(file, ...)
}
\arguments{
\item{file}{If suffix is ".gz" it is assumed the file is gzipped.}

\item{...}{passed to data.table::fread()}
}
\description{
Read a tab-separated file, drop the first column and use the values of the first column as rownames.
}
