% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_util.R
\name{removeSuffix}
\alias{removeSuffix}
\title{Remove a suffix from a string}
\usage{
removeSuffix(str, suffix)
}
\arguments{
\item{str}{the string to be truncated}

\item{suffix}{the string to be checked for and removed}
}
\value{
the string with suffix removed
}
\description{
It is an error if the string does not end with the suffix.
}
