% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_util.R
\name{replaceSuffix}
\alias{replaceSuffix}
\title{Remove a suffix from a file and append a new one.}
\usage{
replaceSuffix(str, oldSuffix, newSuffix)
}
\arguments{
\item{str}{String on which to operate.}

\item{oldSuffix}{suffix to be replaced.}

\item{newSuffix}{the replacement.}
}
\value{
str with oldSuffix replaced by newSuffix
}
\description{
Remove a suffix from a file and append a new one.
}
