% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_util.R
\name{splitStringByLength}
\alias{splitStringByLength}
\title{Split string into pieces of the given length}
\usage{
splitStringByLength(s, len)
}
\arguments{
\item{s}{String to be split}

\item{len}{desired length}
}
\value{
vector of pieces.  Note that last piece may not be full length.
}
\description{
Split string into pieces of the given length
}
