% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_util.R
\name{strEndsWith}
\alias{strEndsWith}
\title{Test if string ends with given suffix}
\usage{
strEndsWith(theString, theExt)
}
\arguments{
\item{theString}{string to be tested}

\item{theExt}{suffix to be tested for}
}
\value{
TRUE if theString ends with theExt
}
\description{
Test if string ends with given suffix
}
