% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_util.R
\name{strStartsWith}
\alias{strStartsWith}
\title{Test if string starts with given prefix}
\usage{
strStartsWith(theString, thePrefix)
}
\arguments{
\item{theString}{string to be tested}

\item{thePrefix}{prefix to be tested for}
}
\value{
TRUE if theString starts with thePrefix
}
\description{
Test if string starts with given prefix
}
