% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_util.R
\name{write_table_with_label_for_rownames}
\alias{write_table_with_label_for_rownames}
\title{Write a tab-separated file with row and column labels, including a label for the rownames.
Additional args are passed to write.table}
\usage{
write_table_with_label_for_rownames(
  x,
  file,
  rowname_label,
  quote = FALSE,
  sep = "\\t",
  col.names = TRUE,
  ...
)
}
\arguments{
\item{x}{the data frame to be written.}

\item{file}{where to write the data.  If suffix is ".gz", the file is gzipped.}

\item{rowname_label}{the column header for the rownames}

\item{quote}{passed to write.table}

\item{sep}{passed to write.table}

\item{col.names}{passed to write.table}

\item{...}{passed to write.table}
}
\description{
Write a tab-separated file with row and column labels, including a label for the rownames.
Additional args are passed to write.table
}
